Wrap all collectors in a CollectorManager
This change removes MergeableCollector (and MergeableCollectorManager),
wraps all of the custom Collector classes in their own CollectorManager,
and removes all remaining occurrences of
CollectorManager.forSequentialExecution from the tests.

This also adds all of the other join collectors, bringing the join
module fully into the CollectorManager club.
