+++ lucene/core/src/java/org/apache/lucene/search/CollectorManager.java
@@ -18,8 +18,6 @@ package org.apache.lucene.search;
 
 import java.io.IOException;
 import java.util.Collection;
-import java.util.concurrent.Executor;
-import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.LeafReaderContext;
 
 /**
@@ -55,36 +53,4 @@ public interface CollectorManager<C extends Collector, T> {
    * called after collection is finished on all provided collectors.
    */
   T reduce(Collection<C> collectors) throws IOException;
-
-  /**
-   * Wrap a provided {@link Collector} with a thin {@code CollectorManager} wrapper for use with
-   * {@link IndexSearcher#search(Query, CollectorManager)} when doing single-threaded searching. The
-   * wrapping {@code CollectorManager} provides no {@link CollectorManager#reduce(Collection)}
-   * implementation, so the wrapped {@code Collector} needs to do all relevant work while
-   * collecting.
-   *
-   * <p>Note: This is only safe to use when {@code IndexSearcher} is created with no executor (see:
-   * {@link IndexSearcher#IndexSearcher(IndexReader, Executor)}).
-   */
-  static <C extends Collector> CollectorManager<C, ?> forSequentialExecution(C in) {
-    return new CollectorManager<C, Void>() {
-      private boolean newCollectorInvoked;
-
-      @Override
-      public C newCollector() {
-        if (newCollectorInvoked) {
-          throw new IllegalStateException(
-              "newCollector should be invoked at most once. Ensure your IndexSearcher has been created without an Executor.");
-        }
-        newCollectorInvoked = true;
-        return in;
-      }
-
-      @Override
-      public Void reduce(Collection<C> collectors) {
-        assert collectors.size() == 1 : "collectors should contain exactly one collector instance";
-        return null;
-      }
-    };
-  }
 }
