+++ lucene/join/src/java/org/apache/lucene/search/join/GlobalOrdinalsCollectorManager.java
@@ -1,0 +1,125 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.lucene.search.join;
+
+import java.io.IOException;
+import java.util.Collection;
+import org.apache.lucene.index.DocValues;
+import org.apache.lucene.index.LeafReaderContext;
+import org.apache.lucene.index.OrdinalMap;
+import org.apache.lucene.index.SortedDocValues;
+import org.apache.lucene.search.Collector;
+import org.apache.lucene.search.CollectorManager;
+import org.apache.lucene.search.LeafCollector;
+import org.apache.lucene.search.Scorable;
+import org.apache.lucene.util.LongBitSet;
+import org.apache.lucene.util.LongValues;
+
+/**
+ * A collector that collects all ordinals from a specified field matching the query.
+ *
+ * @lucene.experimental
+ */
+class GlobalOrdinalsCollectorManager
+    implements CollectorManager<
+        GlobalOrdinalsCollectorManager.GlobalOrdinalsCollector, LongBitSet> {
+  private final String field;
+  private final OrdinalMap ordinalMap;
+  private final long valueCount;
+
+  GlobalOrdinalsCollectorManager(String field, OrdinalMap ordinalMap, long valueCount) {
+    this.field = field;
+    this.ordinalMap = ordinalMap;
+    this.valueCount = valueCount;
+  }
+
+  @Override
+  public GlobalOrdinalsCollector newCollector() throws IOException {
+    return new GlobalOrdinalsCollector();
+  }
+
+  @Override
+  public LongBitSet reduce(Collection<GlobalOrdinalsCollector> collectors) throws IOException {
+    LongBitSet combined = new LongBitSet(valueCount);
+    for (GlobalOrdinalsCollector collector : collectors) {
+      combined.or(collector.collectedOrds);
+    }
+    return combined;
+  }
+
+  final class GlobalOrdinalsCollector implements Collector {
+    private final LongBitSet collectedOrds = new LongBitSet(valueCount);
+
+    @Override
+    public org.apache.lucene.search.ScoreMode scoreMode() {
+      return org.apache.lucene.search.ScoreMode.COMPLETE_NO_SCORES;
+    }
+
+    @Override
+    public LeafCollector getLeafCollector(LeafReaderContext context) throws IOException {
+      SortedDocValues docTermOrds = DocValues.getSorted(context.reader(), field);
+      if (ordinalMap != null) {
+        LongValues segmentOrdToGlobalOrdLookup = ordinalMap.getGlobalOrds(context.ord);
+        return new OrdinalMapCollector(docTermOrds, segmentOrdToGlobalOrdLookup);
+      } else {
+        return new SegmentOrdinalCollector(docTermOrds);
+      }
+    }
+
+    final class OrdinalMapCollector implements LeafCollector {
+
+      private final SortedDocValues docTermOrds;
+      private final LongValues segmentOrdToGlobalOrdLookup;
+
+      OrdinalMapCollector(SortedDocValues docTermOrds, LongValues segmentOrdToGlobalOrdLookup) {
+        this.docTermOrds = docTermOrds;
+        this.segmentOrdToGlobalOrdLookup = segmentOrdToGlobalOrdLookup;
+      }
+
+      @Override
+      public void collect(int doc) throws IOException {
+        if (docTermOrds.advanceExact(doc)) {
+          long segmentOrd = docTermOrds.ordValue();
+          long globalOrd = segmentOrdToGlobalOrdLookup.get(segmentOrd);
+          collectedOrds.set(globalOrd);
+        }
+      }
+
+      @Override
+      public void setScorer(Scorable scorer) throws IOException {}
+    }
+
+    final class SegmentOrdinalCollector implements LeafCollector {
+
+      private final SortedDocValues docTermOrds;
+
+      SegmentOrdinalCollector(SortedDocValues docTermOrds) {
+        this.docTermOrds = docTermOrds;
+      }
+
+      @Override
+      public void collect(int doc) throws IOException {
+        if (docTermOrds.advanceExact(doc)) {
+          collectedOrds.set(docTermOrds.ordValue());
+        }
+      }
+
+      @Override
+      public void setScorer(Scorable scorer) throws IOException {}
+    }
+  }
+}
