+++ lucene/join/src/test/org/apache/lucene/search/join/TestJoinUtil.java
@@ -21,6 +21,7 @@ import com.carrotsearch.randomizedtesting.generators.RandomPicks;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.HashMap;
@@ -32,6 +33,7 @@ import java.util.Random;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
+import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.DoubleDocValuesField;
 import org.apache.lucene.document.DoublePoint;
@@ -466,17 +468,18 @@ public class TestJoinUtil extends LuceneTestCase {
         System.out.println("joinQuery=" + joinQuery);
       }
 
-      final BitSet actualResult = new FixedBitSet(indexSearcher.getIndexReader().maxDoc());
-      final TopScoreDocCollector topScoreDocCollector =
-          new TopScoreDocCollectorManager(10, null, Integer.MAX_VALUE, false).newCollector();
-      newSearcher(indexSearcher.getIndexReader(), true, true, false)
-          .search(
-              joinQuery,
-              CollectorManager.forSequentialExecution(
-                  MultiCollector.wrap(new BitSetCollector(actualResult), topScoreDocCollector)));
+      final BitSetCollectorManager bitsetCollectorManager =
+          new BitSetCollectorManager(indexSearcher.getIndexReader().maxDoc());
+      final TopScoreDocCollectorManager topScoreDocCollectorManager =
+          new TopScoreDocCollectorManager(10, null, Integer.MAX_VALUE);
+      final MultiCollectorManager multiCollectorManager =
+          new MultiCollectorManager(bitsetCollectorManager, topScoreDocCollectorManager);
+      Object[] searchResult = indexSearcher.search(joinQuery, multiCollectorManager);
+      BitSet actualResult = (BitSet) searchResult[0];
+      TopDocs actualTopDocs = (TopDocs) searchResult[1];
+
       assertBitSet(expectedResult, actualResult, indexSearcher);
       TopDocs expectedTopDocs = createExpectedTopDocs(randomValue, from, scoreMode, context);
-      TopDocs actualTopDocs = topScoreDocCollector.topDocs();
       assertTopDocs(expectedTopDocs, actualTopDocs, scoreMode, indexSearcher, joinQuery);
     }
     context.close();
@@ -900,21 +903,23 @@ public class TestJoinUtil extends LuceneTestCase {
 
     indexSearcher.search(
         bq.build(),
-        CollectorManager.forSequentialExecution(
-            new SimpleCollector() {
-              boolean sawFive;
+        new CollectorManager<SimpleCollector, Boolean>() {
+          final AtomicBoolean sawFive = new AtomicBoolean();
 
+          @Override
+          public SimpleCollector newCollector() throws IOException {
+            return new SimpleCollector() {
               @Override
               public void collect(int docID) {
                 // Hairy / evil (depends on how BooleanScorer
                 // stores temporarily collected docIDs by
                 // appending to head of linked list):
                 if (docID == 5) {
-                  sawFive = true;
+                  sawFive.set(true);
                 } else if (docID == 1) {
                   assertFalse(
                       "optimized bulkScorer was not used for join query embedded in boolean query!",
-                      sawFive);
+                      sawFive.get());
                 }
               }
 
@@ -922,7 +927,14 @@ public class TestJoinUtil extends LuceneTestCase {
               public org.apache.lucene.search.ScoreMode scoreMode() {
                 return org.apache.lucene.search.ScoreMode.COMPLETE_NO_SCORES;
               }
-            }));
+            };
+          }
+
+          @Override
+          public Boolean reduce(Collection<SimpleCollector> collectors) throws IOException {
+            return sawFive.get();
+          }
+        });
 
     indexSearcher.getIndexReader().close();
     dir.close();
@@ -1474,9 +1486,7 @@ public class TestJoinUtil extends LuceneTestCase {
       }
       IndexIterationContext context =
           createContext(numberOfDocumentsToIndex, multipleValuesPerDocument, false);
-      // TODO: Once we support concurrency on these join types, don't replace context's searcher
-      IndexSearcher indexSearcher =
-          newSearcher(context.searcher.getIndexReader(), true, true, false);
+      IndexSearcher indexSearcher = context.searcher;
       if (VERBOSE) {
         System.out.println("TEST: got searcher=" + indexSearcher);
       }
@@ -1550,18 +1560,19 @@ public class TestJoinUtil extends LuceneTestCase {
 
         // Need to know all documents that have matches. TopDocs doesn't give me that and then I'd
         // be also testing TopDocsCollector...
-        final BitSet actualResult = new FixedBitSet(indexSearcher.getIndexReader().maxDoc());
-        final TopScoreDocCollector topScoreDocCollector =
-            new TopScoreDocCollectorManager(10, null, Integer.MAX_VALUE, false).newCollector();
-        indexSearcher.search(
-            joinQuery,
-            CollectorManager.forSequentialExecution(
-                MultiCollector.wrap(new BitSetCollector(actualResult), topScoreDocCollector)));
+        final BitSetCollectorManager bitsetCollectorManager =
+            new BitSetCollectorManager(indexSearcher.getIndexReader().maxDoc());
+        final TopScoreDocCollectorManager topScoreDocCollectorManager =
+            new TopScoreDocCollectorManager(10, null, Integer.MAX_VALUE);
+        final MultiCollectorManager multiCollectorManager =
+            new MultiCollectorManager(bitsetCollectorManager, topScoreDocCollectorManager);
+        Object[] searchResult = indexSearcher.search(joinQuery, multiCollectorManager);
+        BitSet actualResult = (BitSet) searchResult[0];
+        TopDocs actualTopDocs = (TopDocs) searchResult[1];
         // Asserting bit set...
         assertBitSet(expectedResult, actualResult, indexSearcher);
         // Asserting TopDocs...
         TopDocs expectedTopDocs = createExpectedTopDocs(randomValue, from, scoreMode, context);
-        TopDocs actualTopDocs = topScoreDocCollector.topDocs();
         assertTopDocs(expectedTopDocs, actualTopDocs, scoreMode, indexSearcher, joinQuery);
       }
       context.close();
@@ -1759,7 +1770,8 @@ public class TestJoinUtil extends LuceneTestCase {
     // possible score for
     // any ScoreMode.
     DirectoryReader topLevelReader = DirectoryReader.open(dir);
-    IndexSearcher searcher = newSearcher(topLevelReader, true, true, false);
+    context.searcher = newSearcher(topLevelReader);
+    IndexSearcher searcher = context.searcher;
     for (int i = 0; i < context.randomUniqueValues.length; i++) {
       String uniqueRandomValue = context.randomUniqueValues[i];
       final String fromField;
@@ -1774,103 +1786,63 @@ public class TestJoinUtil extends LuceneTestCase {
         toField = "from";
         queryVals = context.toHitsToJoinScore;
       }
-      final Map<BytesRef, JoinScore> joinValueToJoinScores = new HashMap<>();
-      if (multipleValuesPerDocument) {
-        searcher.search(
-            new TermQuery(new Term("value", uniqueRandomValue)),
-            CollectorManager.forSequentialExecution(
-                new SimpleCollector() {
+      final Map<BytesRef, JoinScore> joinValueToJoinScores =
+          Collections.synchronizedMap(new HashMap<>());
+      searcher.search(
+          new TermQuery(new Term("value", uniqueRandomValue)),
+          new CollectorManager<SimpleCollector, Void>() {
 
-                  private Scorable scorer;
-                  private SortedSetDocValues docTermOrds;
+            @Override
+            public SimpleCollector newCollector() throws IOException {
+              return new SimpleCollector() {
 
-                  @Override
-                  public void collect(int doc) throws IOException {
-                    if (doc > docTermOrds.docID()) {
-                      docTermOrds.advance(doc);
-                    }
-                    if (doc == docTermOrds.docID()) {
-                      long ord;
-                      for (int j = 0; j < docTermOrds.docValueCount(); j++) {
-                        ord = docTermOrds.nextOrd();
-                        final BytesRef joinValue = docTermOrds.lookupOrd(ord);
-                        JoinScore joinScore = joinValueToJoinScores.get(joinValue);
-                        if (joinScore == null) {
-                          joinValueToJoinScores.put(
-                              BytesRef.deepCopyOf(joinValue), joinScore = new JoinScore());
-                        }
-                        joinScore.addScore(scorer.score());
+                private Scorable scorer;
+                private SortedSetDocValues docTermOrds;
+
+                @Override
+                public void collect(int doc) throws IOException {
+                  if (doc > docTermOrds.docID()) {
+                    docTermOrds.advance(doc);
+                  }
+                  if (doc == docTermOrds.docID()) {
+                    long ord;
+                    for (int j = 0; j < docTermOrds.docValueCount(); j++) {
+                      ord = docTermOrds.nextOrd();
+                      final BytesRef joinValue = docTermOrds.lookupOrd(ord);
+                      JoinScore joinScore = joinValueToJoinScores.get(joinValue);
+                      if (joinScore == null) {
+                        joinValueToJoinScores.put(
+                            BytesRef.deepCopyOf(joinValue), joinScore = new JoinScore());
                       }
+                      joinScore.addScore(scorer.score());
                     }
                   }
+                }
 
-                  @Override
-                  protected void doSetNextReader(LeafReaderContext context) throws IOException {
-                    docTermOrds = DocValues.getSortedSet(context.reader(), fromField);
-                  }
+                @Override
+                protected void doSetNextReader(LeafReaderContext context) throws IOException {
+                  docTermOrds = DocValues.getSortedSet(context.reader(), fromField);
+                }
 
-                  @Override
-                  public void setScorer(Scorable scorer) {
-                    this.scorer = scorer;
-                  }
+                @Override
+                public void setScorer(Scorable scorer) {
+                  this.scorer = scorer;
+                }
 
-                  @Override
-                  public org.apache.lucene.search.ScoreMode scoreMode() {
-                    return org.apache.lucene.search.ScoreMode.COMPLETE;
-                  }
-                }));
-      } else {
-        searcher.search(
-            new TermQuery(new Term("value", uniqueRandomValue)),
-            CollectorManager.forSequentialExecution(
-                new SimpleCollector() {
+                @Override
+                public org.apache.lucene.search.ScoreMode scoreMode() {
+                  return org.apache.lucene.search.ScoreMode.COMPLETE;
+                }
+              };
+            }
 
-                  private Scorable scorer;
-                  private SortedDocValues terms;
+            @Override
+            public Void reduce(Collection<SimpleCollector> collectors) throws IOException {
+              return null;
+            }
+          });
 
-                  @Override
-                  public void collect(int doc) throws IOException {
-                    final BytesRef joinValue;
-                    if (terms.advanceExact(doc)) {
-                      joinValue = terms.lookupOrd(terms.ordValue());
-                    } else {
-                      // missing;
-                      return;
-                    }
-
-                    JoinScore joinScore = joinValueToJoinScores.get(joinValue);
-                    if (joinScore == null) {
-                      joinValueToJoinScores.put(
-                          BytesRef.deepCopyOf(joinValue), joinScore = new JoinScore());
-                    }
-                    if (VERBOSE) {
-                      System.out.println(
-                          "expected val="
-                              + joinValue.utf8ToString()
-                              + " expected score="
-                              + scorer.score());
-                    }
-                    joinScore.addScore(scorer.score());
-                  }
-
-                  @Override
-                  protected void doSetNextReader(LeafReaderContext context) throws IOException {
-                    terms = DocValues.getSorted(context.reader(), fromField);
-                  }
-
-                  @Override
-                  public void setScorer(Scorable scorer) {
-                    this.scorer = scorer;
-                  }
-
-                  @Override
-                  public org.apache.lucene.search.ScoreMode scoreMode() {
-                    return org.apache.lucene.search.ScoreMode.COMPLETE;
-                  }
-                }));
-      }
-
-      final Map<Integer, JoinScore> docToJoinScore = new HashMap<>();
+      final Map<Integer, JoinScore> docToJoinScore = Collections.synchronizedMap(new HashMap<>());
       if (multipleValuesPerDocument) {
         Terms terms = MultiTerms.getTerms(topLevelReader, toField);
         if (terms != null) {
@@ -1897,8 +1869,10 @@ public class TestJoinUtil extends LuceneTestCase {
       } else {
         searcher.search(
             new MatchAllDocsQuery(),
-            CollectorManager.forSequentialExecution(
-                new SimpleCollector() {
+            new CollectorManager<SimpleCollector, Void>() {
+              @Override
+              public SimpleCollector newCollector() throws IOException {
+                return new SimpleCollector() {
 
                   private SortedDocValues terms;
                   private int docBase;
@@ -1932,7 +1906,14 @@ public class TestJoinUtil extends LuceneTestCase {
                   public org.apache.lucene.search.ScoreMode scoreMode() {
                     return org.apache.lucene.search.ScoreMode.COMPLETE_NO_SCORES;
                   }
-                }));
+                };
+              }
+
+              @Override
+              public Void reduce(Collection<SimpleCollector> collectors) throws IOException {
+                return null;
+              }
+            });
       }
       queryVals.put(uniqueRandomValue, docToJoinScore);
     }
@@ -1945,7 +1926,6 @@ public class TestJoinUtil extends LuceneTestCase {
       context.ordinalMap = OrdinalMap.build(null, values, PackedInts.DEFAULT);
     }
 
-    context.searcher = newSearcher(topLevelReader);
     context.dir = dir;
     return context;
   }
@@ -2122,12 +2102,34 @@ public class TestJoinUtil extends LuceneTestCase {
     }
   }
 
+  private static class BitSetCollectorManager implements CollectorManager<BitSetCollector, BitSet> {
+    private final int size;
+
+    private BitSetCollectorManager(int size) {
+      this.size = size;
+    }
+
+    @Override
+    public BitSetCollector newCollector() throws IOException {
+      return new BitSetCollector(new FixedBitSet(size));
+    }
+
+    @Override
+    public BitSet reduce(Collection<BitSetCollector> collectors) throws IOException {
+      FixedBitSet reducedBitset = new FixedBitSet(size);
+      for (BitSetCollector collector : collectors) {
+        reducedBitset.or(collector.bitSet);
+      }
+      return reducedBitset;
+    }
+  }
+
   private static class BitSetCollector extends SimpleCollector {
 
-    private final BitSet bitSet;
+    private final FixedBitSet bitSet;
     private int docBase;
 
-    private BitSetCollector(BitSet bitSet) {
+    private BitSetCollector(FixedBitSet bitSet) {
       this.bitSet = bitSet;
     }
